Refactor datacompiler
- Split out ldutil and timeutil modules
- Add `--dataset-iri` and `--union-file` flags
